ARM: entry: allow ARM-private syscalls to be restarted
System calls will only be restarted after signal handling if they
(a) return an error code indicating that a restart is required and
(b) have `why' set to a non-zero value, to indicate that the signal interrupted them.

This patch leaves `why' set to a non-zero value for ARM-private syscalls,
and only zeroes it for syscalls that are not implemented.

Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>